requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-toolbelt==0.0.1
djangorestframework==3.2.3
gunicorn==19.3.0
Pillow==2.9.0
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
